# anms-core/integration_test/openapi.json
{
  "openapi": "3.0.2",
  "info": {
      "title": "AMMOS-ANMS",
      "description": "AMMOS-ANMS Docs",
      "version": "0.1.0"
  },
  "paths": {
...
      "/ari": {
          "get": {
              "tags": [
                  "ARIs"
              ],
              "summary": "Paged Ari",
              "operationId": "paged_ARI_ari_get",
              "parameters": [
                  {
                      "required": false,
                      "schema": {
                          "title": "Page",
                          "minimum": 1.0,
                          "type": "integer",
                          "default": 1
                      },
                      "name": "page",
                      "in": "query"
                  },
                  {
                      "required": false,
                      "schema": {
                          "title": "Size",
                          "maximum": 100.0,
                          "minimum": 1.0,
                          "type": "integer",
                          "default": 50
                      },
                      "name": "size",
                      "in": "query"
                  }
              ],
              "responses": {
                  "200": {
                      "description": "Successful Response",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/Page_ARI_"
                              }
                          }
                      }
                  },
                  "422": {
                      "description": "Validation Error",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/HTTPValidationError"
                              }
                          }
                      }
                  }
              }
          }
      },
...
      "/formal_parameter": {
          "get": {
              "tags": [
                  "ARIs"
              ],
              "summary": "Paged Formal Parameter",
              "operationId": "paged_formal_parameter_formal_parameter_get",
              "parameters": [
                  {
                      "required": false,
                      "schema": {
                          "title": "Page",
                          "minimum": 1.0,
                          "type": "integer",
                          "default": 1
                      },
                      "name": "page",
                      "in": "query"
                  },
                  {
                      "required": false,
                      "schema": {
                          "title": "Size",
                          "maximum": 100.0,
                          "minimum": 1.0,
                          "type": "integer",
                          "default": 50
                      },
                      "name": "size",
                      "in": "query"
                  }
              ],
              "responses": {
                  "200": {
                      "description": "Successful Response",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/Page_FormalParameter_"
                              }
                          }
                      }
                  },
                  "422": {
                      "description": "Validation Error",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/HTTPValidationError"
                              }
                          }
                      }
                  }
              }
          }
      },
...
      "/formal_parameter/id/{fp_spec_id}": {
          "get": {
              "tags": [
                  "ARIs"
              ],
              "summary": "Formal Parameter By Id",
              "operationId": "formal_parameter_by_id_formal_parameter_id__fp_spec_id__get",
              "parameters": [
                  {
                      "required": true,
                      "schema": {
                          "title": "Fp Spec Id",
                          "type": "integer"
                      },
                      "name": "fp_spec_id",
                      "in": "path"
                  }
              ],
              "responses": {
                  "200": {
                      "description": "Successful Response",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/FormalParameter"
                              }
                          }
                      }
                  },
                  "422": {
                      "description": "Validation Error",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/HTTPValidationError"
                              }
                          }
                      }
                  }
              }
          }
      },
...
  },
  "components": {
      "schemas": {
          "ARI": {
              "title": "ARI",
              "type": "object",
              "properties": {
                  "obj_metadata_id": {
                      "title": "Obj Metadata Id",
                      "type": "integer"
                  },
                  "obj_name": {
                      "title": "Obj Name",
                      "type": "string"
                  },
                  "adm_name": {
                      "title": "Adm Name",
                      "type": "string"
                  },
                  "type_name": {
                      "title": "Type Name",
                      "type": "string"
                  },
                  "namespace_id": {
                      "title": "Namespace Id",
                      "type": "integer"
                  },
                  "obj_id": {
                      "title": "Obj Id",
                      "type": "integer"
                  },
                  "parm_id": {
                      "title": "Parm Id",
                      "type": "integer"
                  },
                  "actual": {
                      "title": "Actual",
                      "type": "boolean"
                  },
                  "display": {
                      "title": "Display",
                      "type": "string"
                  }
              }
          },
...
          "FormalParameter": {
              "title": "FormalParameter",
              "type": "object",
              "properties": {
                  "fp_spec_id": {
                      "title": "Fp Spec Id",
                      "type": "integer"
                  },
                  "num_parms": {
                      "title": "Num Parms",
                      "type": "integer"
                  },
                  "use_desc": {
                      "title": "Use Desc",
                      "type": "string"
                  },
                  "parm_names": {
                      "title": "Parm Names",
                      "type": "string"
                  },
                  "parm_type_name": {
                      "title": "Parm Type Name",
                      "type": "string"
                  },
                  "parm_value_id": {
                      "title": "Parm Value Id",
                      "type": "string"
                  }
              }
          },
          "HTTPValidationError": {
              "title": "HTTPValidationError",
              "type": "object",
              "properties": {
                  "detail": {
                      "title": "Detail",
                      "type": "array",
                      "items": {
                          "$ref": "#/components/schemas/ValidationError"
                      }
                  }
              }
          },
...
          "Page_ARI_": {
              "title": "Page[ARI]",
              "required": [
                  "items",
                  "total",
                  "page",
                  "size"
              ],
              "type": "object",
              "properties": {
                  "items": {
                      "title": "Items",
                      "type": "array",
                      "items": {
                          "$ref": "#/components/schemas/ARI"
                      }
                  },
                  "total": {
                      "title": "Total",
                      "minimum": 0.0,
                      "type": "integer"
                  },
                  "page": {
                      "title": "Page",
                      "minimum": 1.0,
                      "type": "integer"
                  },
                  "size": {
                      "title": "Size",
                      "minimum": 1.0,
                      "type": "integer"
                  }
              }
          },
...
          "Page_FormalParameter_": {
              "title": "Page[FormalParameter]",
              "required": [
                  "items",
                  "total",
                  "page",
                  "size"
              ],
              "type": "object",
              "properties": {
                  "items": {
                      "title": "Items",
                      "type": "array",
                      "items": {
                          "$ref": "#/components/schemas/FormalParameter"
                      }
                  },
                  "total": {
                      "title": "Total",
                      "minimum": 0.0,
                      "type": "integer"
                  },
                  "page": {
                      "title": "Page",
                      "minimum": 1.0,
                      "type": "integer"
                  },
                  "size": {
                      "title": "Size",
                      "minimum": 1.0,
                      "type": "integer"
                  }
              }
          },
...
          "ValidationError": {
              "title": "ValidationError",
              "required": [
                  "loc",
                  "msg",
                  "type"
              ],
              "type": "object",
              "properties": {
                  "loc": {
                      "title": "Location",
                      "type": "array",
                      "items": {
                          "anyOf": [
                              {
                                  "type": "string"
                              },
                              {
                                  "type": "integer"
                              }
                          ]
                      }
                  },
                  "msg": {
                      "title": "Message",
                      "type": "string"
                  },
                  "type": {
                      "title": "Error Type",
                      "type": "string"
                  }
              }
          },
...
      }
  }
}